Added pyxrt linux bindings from @skalade
Removing windows python bindings

Remove check from linux for now

Dropping down to python3.8 for mlir-aie compatibility in the same env. Still having issues not being able to find the CU name in the xclbin when loading the application

commented out state check as it is not working with linux driver (different pyxrt bindings needed?)

Call the script to run unregistered xclbins before in the apprunner init

Copy headerfiles one by one, on linux subprocess isn't liking the wildcards.

Now checks to see if it is a windows filepath rather that just running on windows when doing path translations.

Added a simple test to test various path types.

Getting tests up and running on linux

Changes to arguments to get the latest NPU linux driver working

Rebuilding all the prebuilt applications with the patched version of MLIR-AIE for linux

Stopping nputop erroring out for now

Attempting to bump linux support to py3.10

bumping pyxrt to 310 bindings (thanks Sarunas)

Falling back to older driver version for RyzenAI SW Linux tests

Constraining the appropriate python version based on platform.

Signing does not need to happen on windows based flows, so adding a check. Also, keeping the error status check after an apprun for windows but this does not currently work with the bindings on linux.

Reintroducing the windows python bindings alongside the linux one

Added docker setup for Riallto Linux

Keep stability check in for Windows

Case matters for linux which caught this issue.

+++ tests/test_applications.py
@@ -286,12 +286,13 @@ def _test_appbuild(app, *args):
 
 def check_npu():
     """Utility to check that the IPU is available before a runtime test runs, otherwise skip test."""
-    if not get_device_status() == "OK":
-        pytest.skip('Skipping test because the IPU device is not enabled on this device.')
-    xbu = XBUtil()
-    for app in xbu.list_apps():
-        if app.endswith("IPURiallto"):
-            pytest.skip('Skipping test because the IPU is in an unstable state.')
+    if os.name == "nt":
+        if not get_device_status() == "OK":
+            pytest.skip('Skipping test because the IPU device is not enabled on this device.')
+        xbu = XBUtil()
+        for app in xbu.list_apps():
+            if app.endswith("IPURiallto"):
+                pytest.skip('Skipping test because the IPU is in an unstable state.')
 
 
 def reset_device():
@@ -313,15 +314,16 @@ def manage_testing(request):
     Also save all the test specific artifacts to a given location."""
     yield
 
-    test_name = request.node.name
-    os.makedirs(Path('logs'), exist_ok=True)
-    logdir = Path(f'logs/{test_name}')
-    if os.path.exists(logdir) and os.path.isdir(logdir):
-        shutil.rmtree(logdir)
-    os.makedirs(logdir, exist_ok=True)
-    for f in Path.cwd().glob('*.xclbin'):
-        shutil.move(f, logdir)
-    for f in Path.cwd().glob('*.seq'):
-        shutil.move(f, logdir)
-    for f in Path.cwd().glob('*.mlir'):
-        shutil.move(f, logdir)
+    if os.name == "nt":
+        test_name = request.node.name
+        os.makedirs(Path('logs'), exist_ok=True)
+        logdir = Path(f'logs/{test_name}')
+        if os.path.exists(logdir) and os.path.isdir(logdir):
+            shutil.rmtree(logdir)
+        os.makedirs(logdir, exist_ok=True)
+        for f in Path.cwd().glob('*.xclbin'):
+            shutil.move(f, logdir)
+        for f in Path.cwd().glob('*.seq'):
+            shutil.move(f, logdir)
+        for f in Path.cwd().glob('*.mlir'):
+            shutil.move(f, logdir)
